Heavy code refactoring. Normalized types as much as possible, specifcally: switched game sizes to uint16_t. Extracted the Cell struct that was static in game.c to its own file cell.c so it can have unit tests. Added some unit tests for cell (still missing cases). Removed legacy unfinished brute force solver. Added "cross_io.h", to fix warnings about unsafe I/O functions for MSVC. Refactored the draw_game functions from net_display and net_text to a single file "draw_game.c"

+++ src/include/game_rand.h
@@ -11,6 +11,6 @@
  *pieces
  * @return the randomly generated game
  **/
-game random_game_ext(int width, int height, bool swap, bool allow_cross);
+game random_game_ext(uint16_t width, uint16_t height, bool swap, bool allow_cross);
 
 #endif  // GAME_RAND_H
